change decimal places of model fields

+++ trade_perf/models.py
@@ -3,4 +3,4 @@ from django.db import models
 # Create your models here.
 class PriceData(models.Model):
     date = models.DateField()
-    pnl = models.FloatField()
+    pnl = models.DecimalField(max_digits=9, decimal_places=2)
